Refactor all file-access operations except file tree to use error code rather than exceptions
Since we want the software to be extremely reliable when looking at files
I think a better approach to viewing files is using error codes, and handling those
rather than throwing an exception. Exceptions in file systems can be rather common,
especially when you have drives that have faulted, or use not-so-trusty network shares.